ceo_ext/calculator.py
import sympy
from ceo import ability


@ability
def calculator(expr: str, *args, **kwargs) -> float | str:
    expr = expr.replace(',', '')
    expr = expr.replace('_', '')
    try:
        try:
            return sympy.simplify(expr, rational=None)
        except ValueError as ve:
            return ve.__repr__()
    except sympy.SympifyError as se:
        return se.__repr__()
